run om low quality and fixed zero transitions bug

+++ cityslam/transitions/transitions.py
@@ -77,52 +77,59 @@ def predictions_to_scenes(predictions: np.ndarray, threshold: float = 0.5):
 
         return np.array(scenes, dtype=np.int32)
 
-def add_max_min_cuts(video_file_path, max_scene_length, min_scene_length, cut_file, fps=2):
-    probe = ffmpeg.probe(video_file_path)
-    video_info = next(s for s in probe['streams'] if s['codec_type'] == 'video')
-    video_fps = float(video_info['r_frame_rate'].split('/')[0])
-    
-    # Sometimes the frame rate is given as frames per 1000 seconds?
-    if video_fps / 1000 > 1.0:
-        video_fps = video_fps / 1000
-    
-    print(f"video fps: {video_fps}")
-    max_frames = int(video_fps*max_scene_length)
-    min_frames = int(video_fps*min_scene_length)
-
-    fps_ratio = video_fps / fps
-
-    """
-    we drop too short scenes,
-    for the too long scenes we want to divde them up: 
-    [[scene[0], scene[0] + max_frames - 1],
-        [scene[0] + max_frames, scene[0] + 2*max_frames - 1],
-    ...
-        [scene[0] + (whole_maxes - 1)*max_frames, scene[0] + (whole_maxes)*max_frames - 1],
-        [scene[0] + (whole_maxes)*max_frames, scene[1]]
-    """    
-
-    new_scenes = []
-    with open(cut_file, 'r') as file:
-        scenes = np.loadtxt(file, dtype=int)
-        for i, (scene_start, scene_end) in enumerate(scenes):
-            scene_length = scene_end - scene_start
-            whole_maxes = int(scene_length // max_frames)
-            rest_maxes = int(scene_length % max_frames)
-
-            if scene_length > min_frames:
-                
-                for j in range(whole_maxes):
-                    transition = j == 0 and i != 0
-                    new_scenes.append([int((scene_start + j*max_frames) // fps_ratio), int((scene_start + (j+1)*max_frames - 1) // fps_ratio), transition])
-
-                if rest_maxes >= min_frames:
-                    transition = whole_maxes == 0
-                    new_scenes.append([int((scene_start + (whole_maxes)*max_frames) // fps_ratio), int(scene_end // fps_ratio), transition])
+def add_max_min_cuts(video_file_path, max_scene_length, min_scene_length, cut_file, fps=2, overwrite=False):
+    if not Path(f"{cut_file.parent / cut_file.stem}_cropped.txt").is_file() or overwrite:
+        probe = ffmpeg.probe(video_file_path)
+        video_info = next(s for s in probe['streams'] if s['codec_type'] == 'video')
+        video_fps = float(video_info['r_frame_rate'].split('/')[0])
         
-    with open(f"{cut_file.parent / cut_file.stem}_cropped.txt", 'w') as file:
-            np.savetxt(file, new_scenes, fmt="%d")
+        # Sometimes the frame rate is given as frames per 1000 seconds?
+        if video_fps / 1000 > 1.0:
+            video_fps = video_fps / 1000
+        
+        print(f"video fps: {video_fps}")
+        max_frames = int(video_fps*max_scene_length)
+        min_frames = int(video_fps*min_scene_length)
 
+        fps_ratio = video_fps / fps
+
+        """
+        we drop too short scenes,
+        for the too long scenes we want to divde them up: 
+        [[scene[0], scene[0] + max_frames - 1],
+            [scene[0] + max_frames, scene[0] + 2*max_frames - 1],
+        ...
+            [scene[0] + (whole_maxes - 1)*max_frames, scene[0] + (whole_maxes)*max_frames - 1],
+            [scene[0] + (whole_maxes)*max_frames, scene[1]]
+        """    
+
+        new_scenes = []
+        try:
+            with open(cut_file, 'r') as file:
+                scenes = np.loadtxt(file, dtype=int)
+                scenes = np.atleast_2d(scenes)     
+                for i, (scene_start, scene_end) in enumerate(scenes):
+                    scene_length = scene_end - scene_start
+                    whole_maxes = int(scene_length // max_frames)
+                    rest_maxes = int(scene_length % max_frames)
+
+                    if scene_length > min_frames:
+                        
+                        for j in range(whole_maxes):
+                            transition = j == 0 and i != 0
+                            new_scenes.append([int((scene_start + j*max_frames) // fps_ratio), int((scene_start + (j+1)*max_frames - 1) // fps_ratio), transition])
+
+                        if rest_maxes >= min_frames:
+                            transition = whole_maxes == 0
+                            new_scenes.append([int((scene_start + (whole_maxes)*max_frames) // fps_ratio), int(scene_end // fps_ratio), transition])
+                
+            with open(f"{cut_file.parent / cut_file.stem}_cropped.txt", 'w') as file:
+                    np.savetxt(file, new_scenes, fmt="%d")
+        except Exception as e:
+            print(e)
+    
+    else:
+        print(f"Already found cropped transitions for video {video_file_path.stem}")
 
 def main(videos_dir, video_ids, model_path, output, max_scene_length, min_scene_length, fps, threshold, overwrite=False):
     videos_dir = Path(videos_dir)
@@ -141,7 +148,7 @@ def main(videos_dir, video_ids, model_path, output, max_scene_length, min_scene_
         video_ids = [file[file.find("[")+1:file.find("]")] for file in videos_dir.iterdir()]
 
     for video_id in video_ids:
-
+                    
         video_file_path = next(videos_dir.glob(f"*{video_id}*"))
         if not video_file_path:
             print(f"could not find video with id {video_id}, skipping")
@@ -151,10 +158,18 @@ def main(videos_dir, video_ids, model_path, output, max_scene_length, min_scene_
         output_file.parent.mkdir(exist_ok=True, parents=True)
         
         if not output_file.exists() or overwrite:
-
-            video_stream, err = ffmpeg.input(video_file_path).output(
-                    "pipe:", format="rawvideo", pix_fmt="rgb24", s="48x27"
-                ).run(capture_stdout=True, capture_stderr=True)
+            
+            try:
+                video_stream, err = ffmpeg.input(video_file_path).output(
+                        "pipe:", format="rawvideo", pix_fmt="rgb24", s="48x27"
+                    ).run(capture_stdout=True, capture_stderr=True)
+            except ffmpeg.Error as e:
+                print(video_file_path)
+                #print('stdout:', e.stdout.decode('utf8'))
+                #print('stderr:', e.stderr.decode('utf8'))
+                #raise e
+                #doesent work for some videos for some reason. fex. fGxbtg1ytJo
+                continue   
 
             video = np.frombuffer(video_stream, np.uint8).reshape([-1, 27, 48, 3])
 
@@ -168,13 +183,13 @@ def main(videos_dir, video_ids, model_path, output, max_scene_length, min_scene_
 
         else:
             print(f"Already found transitions for video {video_file_path.stem}")
-
-        add_max_min_cuts(video_file_path, max_scene_length, min_scene_length, output_file, fps)
+        
+        add_max_min_cuts(video_file_path, max_scene_length, min_scene_length, output_file, fps, overwrite) 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument('--videos_dir', type=Path,
-                        default='/cluster/home/ksteinsland/zuricityslam/base/kriss/datasets/videos/🇨🇭 Zürich Switzerland Walk 4K 🌁 4K Walking Tour ☁️ (Cloudy Day) 🇨🇭 [W25QdyiFnh0].mp4',
+                        default='/cluster/project/infk/courses/252-0579-00L/group07/datasets/videos_wv',
                         help='path to videos')
     parser.add_argument('--video_ids', type=str, nargs="+",
                             default=None,
@@ -183,7 +198,7 @@ if __name__ == "__main__":
                         default='./', #'/cluster/project/infk/courses/252-0579-00L/group07/data/dev/TransNetV2/inference/transnetv2-weights',
                         help='path to transiton detection model weights')
     parser.add_argument('--output', type=Path,
-                        default='/cluster/project/infk/courses/252-0579-00L/group07/kriss/datasets/cuts',
+                        default='/cluster/project/infk/courses/252-0579-00L/group07/datasets/transitions',
                         help='where to store list of cuts for specific video')
     parser.add_argument('--max_scene_length', type=int,
                         default=5*60,
@@ -197,7 +212,8 @@ if __name__ == "__main__":
     parser.add_argument('--threshold', type=float,
                         default=0.5,
                         help='transition threshold')
-    parser.add_argument('--overwrite', action="store_true")
+    parser.add_argument('--overwrite', action="store_true",
+                        default=False)                
 
     args = parser.parse_args()
     main(**args.__dict__)
